// SutraPlusUI-main/src/app/admin/components/sales/models/bill-summary.model.ts
export class BillSummary {
    companyId: number;
    ledgerId: number;
    vochType: number;
    vochNo: number;
    displayinvNo: string;
    tranctDate: Date;
    ponumber: string;
    stateCode2: string;
    cessValue: number;
    csgstValue: number;
    igstValue: number;
    sgstValue: number;
    billAmount: number;
    taxableValue: number;
    expenseAmount1: number;
    expenseAmount2: number;
    expenseAmount3: number;
    tcsValue: number;
    advance: number;
    totalAmount: number;
    roundOff: number;
    isSEZ: number;
    shipBillNo: string;
    portName: string;
    shipBillDate: Date;
    dispatcherName: string;
    dispatcherAddress1: string;
    dispatcherAddress2: string;
    dispatcherPlace: string;
    dispatcherPIN: string;
    dispatcherStatecode: string;
    stateCode1: string;
    transporter: string;
    distance: number;
    lorryNo: string;
    deliveryName: string;
    deliveryAddress1: string;
    deliveryAddress2: string;
    deliveryPlace: string;
    delPinCode: string;
    totalWeight: number;
    partyInvoiceNumber: string;
    deliveryStateCode: string;
    IsServiceInvoice:boolean;
  }